comments addressed

+++ include/aws/http/connection.h
@@ -558,11 +558,16 @@ int aws_http2_connection_get_received_goaway(
  * The connection maintains its flow-control windows such that
  * no back-pressure is applied and data arrives as fast as possible.
  *
+ * If you are not connected, this call will have no effect.
+ *
+ * Fails when the input is invalid. The limit of the Maximum Size is 2147483647. If the increment size cause the
+ * connection flow window exceeds the Maximun size, this call will fail.
+ *
  * @param http2_connection HTTP/2 connection.
  * @param increment_size The size to increment for the connection's flow control window
  */
 AWS_HTTP_API
-int aws_http2_connection_update_window(struct aws_http_connection *http2_connection, size_t increment_size);
+int aws_http2_connection_update_window(struct aws_http_connection *http2_connection, uint32_t increment_size);
 
 AWS_EXTERN_C_END
